Add Stage 4/5 project
Now the user is prompted to enter their name, and then the program checks whether there is a record for the user with the same name in a rating.txt file; if yes, then the program uses their score as a starting point for calculating the score in the current game. The user can find out their score by typing '!rating'.